jvb_backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib[bcrypt]
pyjwt
python-dotenv
redis